fixed double

+++ src/parser/DataParser.java
@@ -34,10 +34,10 @@ public class DataParser {
 			Walk walk = new Walk(walker);
 			while ((nextLine = reader.readNext()) != null) {
 				// nextLine[] is an array of values from the line
-				double time= Integer.parseInt(nextLine[0]);
-				double x = Integer.parseInt(nextLine[1]]);
-				double y = Integer.parseInt(nextLine[2]);
-				double z = Integer.parseInt(nextLine[3]);
+				double time= Double.parseDouble(nextLine[0]);
+				double x = Double.parseDouble(nextLine[1]]);
+				double y = Double.parseDouble(nextLine[2]);
+				double z = Double.parseDouble(nextLine[3]);
 				Point point = new Point(time, x,y,z);
 				walk.addPoint(point);
